Move page-router to components/core
