+++ src/components/lineup-view-team-create.ts
@@ -5,10 +5,10 @@ import { html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import { ConnectStoreMixin } from '../middleware/connect-mixin.js';
 import { addNewTeam, getTeamSliceConfigurator } from '../slices/team/index.js';
+import { PageRouter, pageRouterContext } from './core/page-router.js';
 import { AuthorizedViewElement } from './core/page-view-element.js';
 import './lineup-team-create.js';
 import { NewTeamCreatedEvent } from './lineup-team-create.js';
-import { PageRouter, pageRouterContext } from './page-router.js';
 import { SharedStyles } from './shared-styles.js';
 import { SignedInAuthController } from './util/auth-controller.js';
