+++ test/unit/components/lineup-game-setup.test.ts
@@ -1,12 +1,12 @@
 /** @format */
 
 import { RootState, setupStore } from '@app/app/store.js';
+import { PageRouter } from '@app/components/core/page-router.js';
 import '@app/components/lineup-game-setup.js';
 import { LineupGameSetup } from '@app/components/lineup-game-setup.js';
 import { LineupOnPlayerList } from '@app/components/lineup-on-player-list.js';
 import { LineupPlayerCard, PositionSelectedEvent } from '@app/components/lineup-player-card.js';
 import { LineupPlayerList } from '@app/components/lineup-player-list.js';
-import { PageRouter } from '@app/components/page-router.js';
 import { FormationBuilder, FormationType, getPositions } from '@app/models/formation.js';
 import { GameDetail, GameStatus } from '@app/models/game.js';
 import {
